requirements.txt
streamlit 
tiktoken
verovio
transformers
Pillow
numpy
torch
torchvision
accelerate
